+++ packages/user-service/src/index.ts
@@ -34,6 +34,22 @@ const getUserInfoFromCommand = async function() {
   return { account: nick, empId: emp_id };
 }
 
+function promiseWithTimeout<T>(timeoutMs: number, promise: () => Promise<T>, failureMessage?: string) {
+  let timeoutHandle: NodeJS.Timeout;
+  const timeoutPromise = new Promise<never>((resolve, reject) => {
+    timeoutHandle = setTimeout(() => reject(new Error(failureMessage)), timeoutMs);
+  });
+
+  return Promise.race([
+    promise(),
+    timeoutPromise,
+  ]).then((result) => {
+    clearTimeout(timeoutHandle);
+    return result;
+  });
+}
+
+const timeoutMs = 3 * 60 * 1000;
 export async function getUserInfo(): Promise<IUserInfo> {
   // get user info from setting.json
   const userData = configure.get(CONFIGURE_USER_KEY) || {};
@@ -44,7 +60,7 @@ export async function getUserInfo(): Promise<IUserInfo> {
   } else {
     try {
       const fn = isO2 ? getUserInfoFromCommand : getUserInfoFromDefClient;
-      const { account, empid: empId } = await fn();
+      const { account, empid: empId } = await promiseWithTimeout(timeoutMs, fn, 'Timeout!!!');
       const result = { account, empId, gitlabToken };
       configure.set(CONFIGURE_USER_KEY, result);
       return result;